mtl-sequence-tagging-framework/requirements.txt
absl-py==0.10.0
astunparse==1.6.3
boto==2.49.0
boto3==1.14.55
botocore==1.17.55
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
docopt==0.6.2
docutils==0.15.2
editdistance==0.5.3
gast==0.3.3
gensim==3.8.3
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
jmespath==0.10.0
joblib==0.16.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pipreqs==0.4.10
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.4.1
six==1.15.0
sklearn==0.0
smart-open==2.1.1
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-addons==0.11.2
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
typeguard==2.9.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
yarg==0.1.9
zipp==3.1.0
